v0.5.10

+++ src/models/Flight.java
@@ -47,7 +47,7 @@ public class Flight implements Comparable<Flight> {
 		return carrier;
 	}
 
-	public int getMaxPassengers() {
+	public double getMaxPassengers() {
 		return maxPassengers;
 	}
